redo of the order graph, no need to process everything with the notMatching Map.  idk why I didnt think of this initially, but if a namespace doesn't have of all of its imported namespaces in the current OrderGraph then it just shouldn't be processed.  Put it back in the back of the queue and check again later.  Eventually, it will have all of its imports in the order graph and can be placed right after its deepest matching level.
A lot of code was able to be removed because of this switch! Not matching had a lot of processing to it and it was unnecessary.

findNamespacePlaceInOrderGraph takes all the imports of a namespace and finds the levels it matches in.  If it finds one that isn't there, it breaks and returns null, you just can't process it unless ALL imports have a match in the order Graph.