TreeNode edited with some changes.  addChild and addParent have an extra param to pass as a flag, that signals a bi-directional (two-way) relationship with a parent and a child node.
Constructor passed with a list of parents, sets the parents children as
well as setting the parents for the constructed node (child).

Spec tackles these invocations.

TreeNode also contains a getAncestors() method, that climbs up parent
dependencies and returns a set of parent nodes this node depends on.
The spec for this method is not created yet