XsdDependencyTreeFactory is a service object that takes a list of NamespaceData objects and a list of all the BaseSchemaDocument objects collected from the xsd files.
Resolves dependent Namespaces to a map of NamespaceData => Set of
namespaces (strings) depended on.  This is a shallow dependency, only
what is specified in the xsd file is taken into consideration, no deep
dependency resolution.

Creating the dependency tree is simply a matter of figuring out which
dependent namespaces (those namespaces that depend on others) meet
certain criteria (operating on namespace strings associated with
NamespaceData object)

1. at least on depdendency needs to be in the 'current' tree row
2. all depenendcies should be managed already

aquire the namespaces that pass this criteria and send them to the
TreeManager and keep going until there are no more dependent namespaces
left to add to the tree.