Adds args, generatePackage, and task classname options
Adds the ability to specify a number of options. First, you can now
specify the task classname to use so that you can specify the
org.jvnet.jaxb2_commons.xjc.XJC2Task as an option.
Adds the ability to specify arbitrary args to the task, so now
plugins can be activated.
Adds the ability to specify the generatePackage option.
Also sets the verbose/debug flags to the XJCTask based upon the
settings from Gradle.